+++ src/frontend/apps/standalone_site/src/features/Contents/components/Contents/Contents.tsx
@@ -1,5 +1,5 @@
-import { Box, Heading, Text } from 'grommet';
-import { StyledLink } from 'lib-components';
+import { Box, Text } from 'grommet';
+import { Heading, StyledLink } from 'lib-components';
 import { Fragment } from 'react';
 import { defineMessages, useIntl } from 'react-intl';
 import styled from 'styled-components';
@@ -38,7 +38,7 @@ const Contents = ({ playlistId }: ContentsProps) => {
             justify="between"
             margin={{ bottom: 'small' }}
           >
-            <Heading weight="bolder" level="4" margin="none">
+            <Heading level={3} style={{ margin: '0' }}>
               {intl.formatMessage(sample.title)}
             </Heading>
             <Text weight="bolder">