+++ src/frontend/packages/lib_classroom/src/components/ClassroomWidgetProvider/widgets/DeleteClassroom/index.tsx
@@ -1,7 +1,8 @@
-import { Button, Heading, Text } from 'grommet';
+import { Button, Text } from 'grommet';
 import {
   ButtonLoaderStyle,
   FoldableItem,
+  Heading,
   Modal,
   ModalButton,
   report,
@@ -101,11 +102,7 @@ export const DeleteClassroom = () => {
       title={intl.formatMessage(messages.title)}
     >
       <Modal isOpen={isModalOpen} onClose={() => setIsModalOpen(false)}>
-        <Heading
-          size="3"
-          alignSelf="center"
-          margin={{ top: '0', bottom: 'small' }}
-        >
+        <Heading level={2} textAlign="center" className="mt-0 mb-s">
           {intl.formatMessage(messages.deleteModalTitle)}
         </Heading>
         <Text margin={{ top: 'small' }}>